% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alunos.R
\docType{data}
\name{alunos}
\alias{alunos}
\title{Base com dados de alunos}
\format{Uma tabela com 6 colunas e 145264 linhas.
\describe{
  \item{ibge}{codigo ibge do ente federativo}
  \item{nome}{nome do ente federativo}
  \item{regiao}{regiao do ente federativo}
  \item{estado}{estado do ente federativo}
  \item{etapa}{etapa de ensino}
  \item{alunos}{numero de alunos por etapa de ensino}
}}
\source{
Tabela provida por Todos pela educacao
}
\usage{
alunos
}
\description{
Base com dados de alunos de 2015 para o pacote
}
\keyword{datasets}
