% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alunos-teste.R
\docType{data}
\name{alunos_teste}
\alias{alunos_teste}
\title{Base com dados de alunos para teste}
\format{Uma tabela com 3 colunas e 140 linhas.
\describe{
  \item{ibge}{codigo ibge do ente federativo}
  \item{etapa}{etapa de ensino}
  \item{alunos}{numero de alunos por etapa de ensino}
}}
\source{
Tabela criada
}
\usage{
alunos_teste
}
\description{
Base com dados artificiais de alunos para realizacao de teste das funcoes do pacote
}
\keyword{datasets}
