% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula-fundo-estadual.R
\name{calcula_fundo_estadual}
\alias{calcula_fundo_estadual}
\title{Calcula valor do fundo estadual}
\usage{
calcula_fundo_estadual(fundo, codigo = ibge, var_fundo = fundeb)
}
\arguments{
\item{fundo}{data.frame com o valor do fundo de cada ente federativo e o codigo ibge do ente}

\item{codigo}{coluna numerica com o codigo do estado}

\item{var_fundo}{nomeia a coluna numerica com o valor do fundo}
}
\value{
Um data.frame com o valor do fundo estadual por estado
}
\description{
Essa função calcula o valor total dos fundo estaduais, a funcao considera que os dois primeiros digitos do codigo correspondem ao estado
}
\examples{
library(simulador.fundeb)
simulador.fundeb:::calcula_fundo_estadual(test, var_fundo = valor)
}
