% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula-fundo-total.R
\name{calcula_fundo_total}
\alias{calcula_fundo_total}
\title{Calcula valor do fundo total}
\usage{
calcula_fundo_total(fundo, var_fundo = valor)
}
\arguments{
\item{fundo}{data.frame}

\item{var_fundo}{coluna numerica de fundo}
}
\value{
O valor do fundo total como vetor numerico de tamanho 1
}
\description{
Essa função calcula o valor total de um fundo
}
\examples{
library(simulador.fundeb)
simulador.fundeb:::calcula_fundo_total(test)
}
