% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checa_na}
\alias{checa_na}
\title{Checa NAs}
\usage{
checa_na(x, nome)
}
\arguments{
\item{x}{um objeto a ser testado}

\item{nome}{nome do objeto para ser passado para a chamada de erro se necessaria}
}
\value{
Lanca um erro ou retorna um vetor logico
}
\description{
Checa se um objeto possui algum valor NA
}
