% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checa_numerico}
\alias{checa_numerico}
\title{Checa numerico}
\usage{
checa_numerico(x, nome)
}
\arguments{
\item{x}{um objeto a ser testado}

\item{nome}{nome do objeto para ser passado para a chamada de erro se necessaria}
}
\value{
Lanca um erro ou retorna um vetor logico
}
\description{
Testa se objeto e um objeto numerico de dimensao 1
}
