% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checa_transforma_dt}
\alias{checa_transforma_dt}
\title{Checa e caso nao transforma em data.table}
\usage{
checa_transforma_dt(dados)
}
\arguments{
\item{dados}{um objeto data.frame ou data.table}
}
\value{
Um objeto em data.table
}
\description{
Checa se um objeto e data.table e caso nao seja transforma em data.table
}
