% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{complementar}
\alias{complementar}
\title{Dados Teste}
\format{
Uma tabela com 9 colunas e 76 linhas.
\describe{
  \item{ibge}{Variavel numerica com o codigo ibge do ente}
  \item{uf}{Variavel em caractere com o nome da unidades da federacao do ente}
  \item{nome}{Variavel em caracteres com o nome do ente}
  \item{recursos_vaaf}{Variaval numerica receitas do fundeb cada ente federativo utilizadas na etapa VAAF}
  \item{recursos_vaat}{Variaval numerica  com receitas do fundeb cada ente federativo utilizadas na etapa VAAT}
  \item{nse}{Variaval numerica com o indice socioeconomico, no exemplo do teste todos os valores sao um}
  \item{nf}{Variaval numerica com o indice de nivel fiscal, no exemplo do teste todos os valores sao um}
  \item{peso_vaar}{Variavel com a proporcao de complementacao VAAR que o ente recebe}
  \item{inabilitados_vaat}{Variavel logica com o identificador dos entes que nao participarao da complementacao vaat}
  \item{nf}{Variaval com o indice de nivel fiscal, no exemplo do teste todos os valores sao um}
  }
}
\source{
Tabela Criada a partir dos dados do Fundeb, referentes a portaria interministerial 2 de 2024
}
\usage{
complementar
}
\description{
Uma base de dados com as matriculas por etapa, coeficiente de distribuicao, estimativas de receitas e estado por ente federativo
}
\keyword{datasets}
