% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{complementar_teste}
\alias{complementar_teste}
\title{Dados Teste}
\format{

}
\usage{
complementar_teste
}
\description{
Uma base de dados com as matriculas por etapa, coeficiente de distribuicao, estimativas de receitas e estado por ente federativo
}
\keyword{datasets}
