% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equaliza-fundo.R
\name{equaliza_fundo}
\alias{equaliza_fundo}
\title{Equaliza fundo}
\usage{
equaliza_fundo(
  dados,
  complementacao,
  var_ordem,
  var_alunos,
  var_receitas,
  entes_excluidos = NULL,
  produto_dt = TRUE
)
}
\arguments{
\item{dados}{Um objeto da classe data.table ou data.frame com os dados do fundo a serem equalizados}

\item{complementacao}{Numero com a quantidade de fundos a serem utilizados na equalizacao}

\item{var_ordem}{Variavel que sera utilizada para ordenar os entes}

\item{var_alunos}{Variavel que sera utilizada como numero de alunos a serem considerados do ente}

\item{var_receitas}{Variavel com as receitas que seram equalizadas}
}
\value{
Um data.frame ou data.table
}
\description{
Recebe dados dos entes do fnde, divididos em entes federativos ou fundos estaduais, e equaliza o valor recebido de acordo com o vaa em questao. A equalizacao e feita de baixo para cima.
}
