% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equaliza-modelo.R
\name{equaliza_modelo}
\alias{equaliza_modelo}
\title{Equaliza um modelo}
\usage{
equaliza_modelo(
  dados,
  fundo,
  aporte = aporte_federal,
  var_alunos = alunos_estado,
  codigo
)
}
\arguments{
\item{dados}{data.frame que serao equalizados}

\item{fundo}{coluna com o valor do fundo a ser equalizado}

\item{aporte}{quantidade de verba que será usada na equalização}

\item{var_alunos}{coluna numerica com o numero de alunos}

\item{codigo}{codigo do ente ou estado}
}
\value{
Um data.frame com o fundo avaliado equalizado
}
\description{
Função intermediária que equaliza o valor do fundo por aluno
}
\examples{
library(simulador.fundeb)
}
