% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financas.R
\docType{data}
\name{financas}
\alias{financas}
\title{Base com dados de financiamento do FUNDEB}
\format{Uma tabela com 6 colunas e 33552 linhas.
\describe{
  \item{ibge}{codigo ibge do ente federativo}
  \item{nome}{nome do ente federativo}
  \item{regiao}{regiao do ente federativo}
  \item{estado}{estado do ente federativo}
  \item{tipo}{tipo do fundo}
  \item{valor}{valor do fundo}
}}
\source{
Tabela provida por Todos pela educacao
}
\usage{
financas
}
\description{
Base com dados de financiamente do FUNDEB de 2015 para o pacote
}
\keyword{datasets}
