% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financas-teste.R
\docType{data}
\name{financas_teste}
\alias{financas_teste}
\title{Base com dados de financiamento do FUNDEB para teste das funcoes}
\format{
Uma tabela com 3 colunas e 5 linhas.
\describe{
  \item{ibge}{codigo ibge do ente federativo}
  \item{tipo}{tipo do fundo}
  \item{valor}{valor do fundo}
}
}
\source{
Tabela criada
}
\usage{
financas_teste
}
\description{
Base com dados artificiais de financiamente para realizacao de testes das funcoes do pacote
}
\keyword{datasets}
