% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-dados-estaduais.R
\name{gera_dados_estaduais}
\alias{gera_dados_estaduais}
\title{Gera tabela com dados dos estados}
\usage{
gera_dados_estaduais(dados)
}
\arguments{
\item{dados}{data.frame preparado por pondera_geral}
}
\value{
Um data.frame com dados de alunos por estado, pela ponderacao por etapa e por dados socioeconomicos, fundo estadual, vaa, pelas duas ponderacoes de aluno
}
\description{
Funcao intermediaria que agrega dados de ente federativo por estado
}
\examples{
library(simulador.fundeb)
}
