% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera_fundo_estadual.R
\name{gera_fundo_estadual}
\alias{gera_fundo_estadual}
\title{Gera Fundo estadual}
\usage{
gera_fundo_estadual(dados_entes)
}
\value{
Um data.frame com os dados de recursos e alunos dos fundos estaduais
}
\description{
Pondera dados de alunos do FNDE por ente e etapa e retorna os valores ponderados por etapa
}
