% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limpa-fnde.R
\name{limpa_fnde}
\alias{limpa_fnde}
\title{Limpa dados do FNDE}
\usage{
limpa_fnde(dados_fnde, produto_dt = TRUE)
}
\arguments{
\item{dados_fnde}{data.frame ou data.table com os dados do FNDE}

\item{produto_dt}{Variavel binaria que determina se a tabela produto da funcao sera da classe data.table (se Verdadeiro) ou data.frame (se Falso)}
}
\value{
Um data.frame ou data.table com os dados do FNDE
}
\description{
Limpa os dados do FNDE removendo coeficiente de distribuicao e o municipio "TOTAL GERAL"
}
