% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo-fundeb.R
\name{modelo_fundeb}
\alias{modelo_fundeb}
\title{Pondera alunos por etapa de ensino}
\usage{
modelo_fundeb(base_alunos, ponderador, base_socioeconomica, base_financas,
  auxilio_federal = 0.1, var_fundo = fundeb, var_alunos = alunos,
  ...)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{ponderador}{data.frame de peso de aluno por etapa}

\item{base_socioeconomica}{data.frame com  o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame com o codigo do ibge e os valores do fundeb e das demais receitas}

\item{auxilio_federal}{percentual do fundo que a União complementará}

\item{var_fundo}{parametro com o nome da variavel do fundo a ser considerado}

\item{var_alunos}{parametro com o nome da variavel alunos a ser considerada}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com codigo ibge, etapa e alunos e outra com pesos de ponderador por etapa e gera uma tabela com o número de alunos ponderados por entidade da federação
}
