% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{pesos}
\alias{pesos}
\title{Peso}
\format{
Uma tabela com 5 colunas e 29 linhas.
\describe{
  \item{nome}{Variavel em caractere com o nome das etapas formatado para apresentacao}
  \item{etapa}{Variavel em caractere com o nome das etapas segundo as portarias do FNDE}
  \item{peso_vaaf}{Variavel numerica com o peso de cada etapa para a etapa de complementacao VAAF segunda o regulamentacao do FNDE}
  \item{peso_vaat}{Variavel numerica com o peso de cada etapa para a etapa de complementacao VAAT}
  }
}
\source{
Tabela Criada a partir dos dados do Fundeb, referentes a portaria interministerial 2 de 2024
}
\usage{
pesos
}
\description{
Uma base de dados com os pesos por etapa
}
\keyword{datasets}
