% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-alunos-etapa.R
\name{pondera_alunos_etapa}
\alias{pondera_alunos_etapa}
\title{Pondera alunos por etapa de ensino}
\usage{
pondera_alunos_etapa(
  base_alunos,
  ponderador,
  codigo = ibge,
  var_alunos = alunos,
  variavel_peso = peso
)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{ponderador}{data.frame de peso de aluno por etapa}

\item{codigo}{coluna numerica com o codigo do ente federativo ou estado}

\item{var_alunos}{nomeia a coluna numerica com o numero de alunos que sera utilizada}

\item{variavel_peso}{nomeia coluna que sera utilizada como peso da etapa de ensino}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com codigo ibge, etapa e alunos e outra com pesos de ponderador por etapa e gera uma tabela com o numero de alunos ponderados por etapa para cada entidade do grupo considerado
}
\examples{
library(simulador.fundeb)
}
