% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-alunos-rede.R
\name{pondera_alunos_rede}
\alias{pondera_alunos_rede}
\title{Calcula valor do fundo total}
\usage{
pondera_alunos_rede(base_alunos, codigo = ibge, etapa_ensino = etapa,
  var_alunos = alunos, busca_creche = "Creche|Pré-escola",
  busca_em = "EM")
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{ponderador}{data.frame de peso de aluno por etapa}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com código ibge, alunos e
}
\examples{
library(simulador.fundeb)
simulador.fundeb:::pondera_alunos_rede(test)

}
