% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-geral.R
\name{pondera_geral}
\alias{pondera_geral}
\title{Pondera alunos por etapa de ensino}
\usage{
pondera_geral(
  base_alunos,
  ponderador,
  base_socioeconomica,
  base_financas,
  condicao_rede = TRUE,
  min_social = 1,
  max_social = 1.2,
  min_disp_fiscal = 1,
  max_disp_fiscal = 1.2,
  var_socioeconomica = nse,
  considerar = "ambos",
  desconsidera_estados = FALSE,
  ...
)
}
\arguments{
\item{base_alunos}{data.frame com numero de alunos por etapa e ente federativo}

\item{base_socioeconomica}{data.frame com o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame com o codigo do ibge e valores do fundeb e das demais receitas}

\item{condicao_rede}{valor binario que condiciona a ponderacao por rede, se verdadeiro atribui pesos diferentes por rede de ensino, ou seja, difere o peso atribuido aos alunos das redes estadual da rede municipal a depender se estao no ensino infantil ou medio}

\item{min_social}{peso minimo dado a informacao socioeconomica}

\item{max_social}{peso maximo dado a informacao socioeconomica}

\item{min_disp_fiscal}{peso minimo dado a informacao de financas}

\item{max_disp_fiscal}{peso maximo dado a informacao de financas}

\item{var_socioeconomica}{variavel numerica socioeconomica de um ente federativo a ser considerada na ponderacao socioeconomica do fundo}

\item{considerar}{nome que indica que parametro sera considarado na criacao do peso socioeconomico, se selecionado social se considerara apenas as informações sociais, se selecionado financas se considerara apenas as informações de financas e se selecionado ambos se considerara ambas as dimensoes}

\item{desconsidera_estados}{variavel logica que indica se os estados serao considerados na ponderacao socioeconomica, se verdadeiro o valor que os estados receberao nao e alterado pelo seu peso socioeconômico, se falso o valor que os estados receberao e alterado pelo seu peso socioeconômico}
}
\value{
Data.frame com alunos ponderados de acordo com as especificacoes providas pelo usuario, dados de financas e dados socioeconomicos
}
\description{
Recebe uma base com codigo ibge, etapa e alunos e outra com pesos de ponderador por etapa e gera uma tabela com o número de alunos ponderados por entidade da federação
}
\examples{
library(simulador.fundeb)
}
