% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera_alunos_etapa.R
\name{pondera_matriculas_etapa}
\alias{pondera_matriculas_etapa}
\title{Pondera matriculas por etapa}
\usage{
pondera_matriculas_etapa(dados_matriculas, dados_peso)
}
\arguments{
\item{dados_matriculas}{data.frame com os dados de matriculas}

\item{dados_peso}{data.frame com os dados de peso das etapas}
}
\value{
Um data.frame ou data.table com os dados de matriculas considerando os pesos dados de FNDE em formato longo ou curto
}
\description{
Pondera dados de matriculas do FNDE por ente e etapa e retorna os valores ponderados por etapa
}
