% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera_alunos_sociofiscal.R
\name{pondera_matriculas_sociofiscal}
\alias{pondera_matriculas_sociofiscal}
\title{Pondera matriculas por nivelsocioeconomico}
\usage{
pondera_matriculas_sociofiscal(dados_matriculas, dados_complementar)
}
\arguments{
\item{dados_matriculas}{data.frame com os dados de matriculas}

\item{dados_complementar}{data.frame com os dados de peso das etapas}
}
\value{
Um data.frame com os dados de matriculas considerando o nivel socioeconomico
}
\description{
Une os dados de matriculas com os dados complementares e atribui o peso do fator socioeconomico
}
