% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-socioeconomico.R
\name{pondera_socioeconomico}
\alias{pondera_socioeconomico}
\title{Pondera alunos por etapa de ensino}
\usage{
pondera_socioeconomico(
  base_alunos,
  base_socioeconomica,
  base_financas,
  var_socioeconomica = nse,
  var_fundo_pond = demais_receitas,
  min_social = 1,
  max_social = 1.2,
  min_disp_fiscal = 1,
  max_disp_fiscal = 1.2,
  considerar = c("social", "financas", "ambos"),
  desconsidera_estados = FALSE,
  codigo = ibge
)
}
\arguments{
\item{base_alunos}{data.frame com numero de alunos por etapa e ente federativo}

\item{base_socioeconomica}{data.frame com o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame com o codigo do ibge e valores do fundeb e das demais receitas}

\item{var_socioeconomica}{variavel numerica socioeconomica de um ente federativo a ser considerada na ponderacao socioeconomica do fundo}

\item{var_fundo_pond}{variavel numerica com o valor do fundo a ser considerado na ponderacao financeira}

\item{min_social}{peso minimo dado a informacao socioeconomica}

\item{max_social}{peso maximo dado a informacao socioeconomica}

\item{min_disp_fiscal}{peso minimo dado a informacao de financas}

\item{max_disp_fiscal}{peso maximo dado a informacao de financas}

\item{considerar}{nome que indica que parametro sera considarado na criacao do peso socioeconomico, se selecionado social se considerara apenas as informações sociais, se selecionado financas se considerara apenas as informações de financas e se selecionado ambos se considerara ambas as dimensoes}

\item{desconsidera_estados}{variavel logica que indica se os estados serao considerados na ponderacao socioeconomica, se verdadeiro o valor que os estados receberao nao e alterado pelo seu peso socioeconômico, se falso o valor que os estados receberao e alterado pelo seu peso socioeconômico}

\item{codigo}{parametro com o nome do codigo a ser usado como identificador dos entes federativos, recomedanda-se o codigo ibge}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com código ibge, etapa e alunos e outra com pesos de ponderador por etapa e gera uma tabela com o número de alunos ponderados por entidade da federação
}
\examples{
library(simulador.fundeb)


}
