% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderador_teste.R
\docType{data}
\name{ponderador_alunos}
\alias{ponderador_alunos}
\title{Base de teste para o pacote}
\format{Uma tabela com 2 colunas e 28 linhas.
\describe{
  \item{etapa}{etapa de ensino}
  \item{ponderador}{peso do aluno por etapa}
}}
\source{
Tabela provida por Todos pela educacao
}
\usage{
ponderador_alunos
}
\description{
Uma base de dados para testar as funcoes do pacote
}
\keyword{datasets}
