% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepara-equalizacao.R
\name{prepara_equalizacao}
\alias{prepara_equalizacao}
\title{Prepara um modelo para equalizacao}
\usage{
prepara_equalizacao(dados, var_vaa = vaa,
  var_alunos_grupo = alunos_estado)
}
\arguments{
\item{dados}{data.frame que sera preparado para equalizacao}

\item{var_vaa}{variavel de valor as ser equalizada entre entes federativos}

\item{var_alunos_grupos}{coluna numerica com o numero de alunos por ente}
}
\value{
Um data.frame com as seguintes colunas adicionas cumulativo de alunos e gasto necessario para equalizar
}
\description{
Função intermediária que prepara um conjunto de dados para equalização gerando variaveis de interesse intermediárias
}
\examples{
library(simulador.fundeb)
}
