% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reescala_vetor}
\alias{reescala_vetor}
\title{Reescala uma vetor}
\usage{
reescala_vetor(var, teto = 1.3, chao = 1)
}
\arguments{
\item{var}{um vetor numerico}

\item{chao}{valor minimo do vetor apois o rescalonamento}

\item{topo}{valor maximo do vetor apos o rescalonamento}
}
\value{
Um vetor numerico reescalado
}
\description{
Checa se um objeto e data.table e caso nao seja transforma em data.table. Caso seja faz uma copia para nao gerar problemas com avaliacoes intermediarias
}
