% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{retorna_dt_df}
\alias{retorna_dt_df}
\title{Retorna data.frame ou data.table}
\usage{
retorna_dt_df(dados, produto_dt)
}
\arguments{
\item{dados}{Um objeto da classe data.table}
}
\value{
Um data.frame ou data.table
}
\description{
Recebe um data.table e determina se transforma em data.frame ou continua dt
}
