% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simula-fundeb.R
\name{simula_fundeb}
\alias{simula_fundeb}
\title{Simula Fundeb}
\usage{
simula_fundeb(
  dados_alunos,
  dados_complementar,
  peso_etapas = peso,
  chao_socio = 1,
  teto_socio = 1,
  difere_etapas_complementacao = c("vaaf_vaat", "mesmos_pesos"),
  chao_fiscal = 1,
  teto_fiscal = 1,
  entes_excluidos_vaat = NULL,
  complementacao_vaaf,
  complementacao_vaat,
  produto_dt = TRUE
)
}
\arguments{
\item{dados_alunos}{data.frame com os dados de alunos}

\item{dados_complementar}{data.frame com os dados de peso das etapas}

\item{peso_etapas}{data.frame com os dados de peso das etapas}

\item{chao_socio}{valor minimo do peso da variavel social na ponderacao de alunos}

\item{teto_socio}{valor maximo do peso da variavel social na ponderacao de alunos}

\item{difere_etapas_complementacao}{variavel em caractere com as opcoes de diferenciacao dos pesos das etapas. Caso escolha-se vaaf_vaat a etapa vaaf e a etapa vaat consideram pesos diferentes, caso escolha-se mesmos pesos as etapas tem os mesmos pesos}

\item{chao_fiscal}{valor minimo do peso da variavel fiscal na ponderacao de alunos}

\item{teto_fiscal}{valor maximo do peso da variavel fiscal na ponderacao de alunos}

\item{complementacao_vaaf}{valor numerico com o montante a ser complementado pela uniao na etapa VAAF}

\item{complementacao_vaat}{valor numerico com o montante a ser complementado pela uniao na etapa VAAT}

\item{produto_dt}{valor logico que determina se o resultado sera em data.table (caso produto_dt = TRUE) ou data.frame (caso produto_dt = FALSE)}
}
\value{
Um data.frame ou data.table com a simulacao dos dados do FNDE
}
\description{
Simula a distribuição dos fundos do Fundeb e as etapas de equalização
}
