% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular-modelo-vaat.R
\name{simular_modelo_vaat}
\alias{simular_modelo_vaat}
\title{Simulação do modelo VAT de financiamento da educação}
\usage{
simular_modelo_vaat(
  base_alunos,
  ponderador,
  base_socioeconomica,
  base_financas,
  complem_uniao = 0.1,
  var_fundo = fundeb,
  var_alunos = alunos,
  condicao_rede = TRUE,
  fatores_intra_equidade = FALSE,
  equalizacao_socio = FALSE,
  min_social = 1,
  max_social = 1.2,
  min_disp_fiscal = 1,
  max_disp_fiscal = 1.2,
  var_socioeconomica = nse,
  considerar = "ambos",
  desconsidera_estados = FALSE,
  ...
)
}
\arguments{
\item{base_alunos}{data.frame com numero de alunos por etapa e ente federativo}

\item{base_socioeconomica}{data.frame com o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame com o codigo do ibge e valores do fundeb e das demais receitas}

\item{complem_uniao}{percentual do fundo que a União complementará}

\item{condicao_rede}{valor binario que condiciona a ponderacao por rede, se verdadeiro atribui pesos diferentes por rede de ensino, ou seja, difere o peso atribuido aos alunos das redes estadual da rede municipal a depender se estao no ensino infantil ou medio}

\item{fatores_intra_equidade}{parametro logico que controla se a distribuicao do fundo estadual considerara o vetor de alunos ou de alunos socioeconomico, se verdadeiro considera-se o vetor de alunos ponderados pelos pesos socioeconomicos caso falso considera-se o vetor de alunos sem ponderacao}

\item{equalizacao_socio}{parametro lógico que controla se a equalização do fundo considerara o vetor de alunos ou de alunos socioeconomico, se verdadeiro considera-se o vetor de alunos ponderados pelos pesos socioeconomicos caso falso considera-se o vetor de alunos sem ponderacao}

\item{min_social}{peso minimo dado a informacao socioeconomica}

\item{max_social}{peso maximo dado a informacao socioeconomica}

\item{min_disp_fiscal}{peso minimo dado a informacao de financas}

\item{max_disp_fiscal}{peso maximo dado a informacao de financas}

\item{var_socioeconomica}{variavel numerica socioeconomica de um ente federativo a ser considerada na ponderacao socioeconomica do fundo}

\item{considerar}{nome que indica que parametro sera considarado na criacao do peso socioeconomico, se selecionado social se considerara apenas as informacoes sociais, se selecionado financas se considerara apenas as informacoes de financas e se selecionado ambos se considerara ambas as dimensoes}

\item{desconsidera_estados}{variavel logica que indica se os estados serao considerados na ponderacao socioeconomica, se verdadeiro o valor que os estados receberao nao e alterado pelo seu peso socioeconomico, se falso o valor que os estados receberao e alterado pelo seu peso socioeconomico}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com numero de alunos por ente e por etapa, ponderador por etapa, dados socioeconomicos por ente e dados financeiros por ente e simula o modelo fundeb de financiamento da educação
}
