% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular-modelo-vat.R
\name{simular_modelo_vat}
\alias{simular_modelo_vat}
\title{Simulação do modelo VAT de financiamento da educação}
\usage{
simular_modelo_vat(
  base_alunos,
  ponderador,
  base_socioeconomica,
  base_financas,
  auxilio_federal = 0.1,
  var_fundo = fundeb,
  var_alunos = alunos,
  distribuicao_fundo_estadual_socio = FALSE,
  equalizacao_socio = FALSE,
  min_social = 1,
  max_social = 1.3,
  min_financas = 1,
  max_financas = 1.3,
  ...
)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{ponderador}{data.frame de peso de aluno por etapa}

\item{base_socioeconomica}{data.frame com  o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame com o codigo do ibge e os valores do fundeb e das demais receitas}

\item{auxilio_federal}{percentual do fundo que a União complementará}

\item{distribuicao_fundo_estadual_socio}{parametro logico que controla se a distribuicao do fundo estadual considerara o vetor de alunos ou de alunos socioeconomico}

\item{equalizacao_socio}{parametro lógico que controla se a equalização do fundo considerara o vetor de alunos ou de alunos socioeconomico}

\item{min_social}{peso minimo dado a informacao socioeconomica}

\item{max_social}{peso maximo dado a informacao socioeconomica}

\item{min_financas}{peso minimo dado a informacao de financas}

\item{max_financas}{peso maximo dado a informacao de financas}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com numero de alunos por ente e por etapa, ponderador por etapa, dados socioeconomicos por ente e dados financeiros por ente e simula o modelo fundeb de financiamento da educação
}
