% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular-modelo-vat-tempo.R
\name{simular_modelo_vat_tempo}
\alias{simular_modelo_vat_tempo}
\title{Simulação do modelo vat de financiamento da educação no tempo}
\usage{
simular_modelo_vat_tempo(
  base_alunos,
  ponderador,
  base_socioeconomica,
  base_financas,
  auxilio_federal = 0.1,
  equalizacao_socio = FALSE,
  distribuicao_fundo_estadual_socio = FALSE,
  crescimento_economico,
  crescimento_demografico,
  min_social = 1,
  max_social = 1.3,
  min_financas = 1,
  max_financas = 1.3,
  var_socioeconomica = nse,
  considerar = "ambos",
  ...
)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{base_socioeconomica}{data.frame o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame o codigo do ibge e valores do fundeb e das demais receitas}

\item{auxilio_federal}{percentual do fundo que a União complementará}

\item{equalizacao_socio}{parametro lógico que controla se a equalização do fundo considerara o vetor de alunos ou de alunos socioeconomico}

\item{distribuicao_fundo_estadual_socio}{parametro logico que controla se a distribuicao do fundo estadual considerara o vetor de alunos ou de alunos socioeconomico}

\item{min_social}{peso minimo dado a informacao socioeconomica}

\item{max_social}{peso maximo dado a informacao socioeconomica}

\item{min_financas}{peso minimo dado a informacao de financas}

\item{max_financas}{peso maximo dado a informacao de financas}

\item{var_socioeconomica}{variavel numerica socioeconomica de um ente federativo}

\item{considerar}{nome que indica que parametro sera considarado na criacao do peso socioeconomico, se selecionado social se considerara apenas as informacoes sociais, se selecionado financas se considerara apenas as informacoes de financas e se selecionado ambos se considerara ambas as dimensoes}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com numero de alunos por ente e por etapa, ponderador por etapa, dados socioeconomicos por ente e dados financeiros por ente e simula o modelo fundeb de financiamento da educação ao longo do tempo. Considera o crescimento demográfico e econômico ao longo do tempo.
}
