% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeco-2015.R
\docType{data}
\name{socioeco_2015}
\alias{socioeco_2015}
\title{Base com dados de valor socioeconomicos}
\format{
Uma tabela com 6 colunas e 5592 linhas.
\describe{
  \item{ibge}{codigo ibge do ente federativo}
  \item{nome}{nome do ente federativo}
  \item{regiao}{regiao do ente federativo}
  \item{estado}{estado do ente federativo}
  \item{nse}{valor de indicador de nivel socioeconomico}
  \item{idhm}{valor de indice de desenvolvimento humano municipal}
}
}
\source{
Tabela provida por Todos pela educacao
}
\usage{
socioeco_2015
}
\description{
Base com dados de financiamento de 2015
}
\keyword{datasets}
