% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socioeco-teste.R
\docType{data}
\name{socioeco_teste}
\alias{socioeco_teste}
\title{Base com dados de valor socioeconomicos inventados}
\format{Uma tabela com 3 colunas e 5 linhas.
\describe{
  \item{ibge}{codigo ibge do ente federativo}
  \item{nse}{valor de indicador de nivel socioeconomico}
  \item{idhm}{valor de indice de desenvolvimento humano municipal}
}}
\source{
Tabela criada
}
\usage{
socioeco_teste
}
\description{
Base com dados dde financiamento por ente, inventada
}
\keyword{datasets}
