% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\docType{data}
\name{test}
\alias{test}
\title{Base de teste para o pacote}
\format{Uma tabela com 3 colunas e 7 linhas.
\describe{
  \item{ibge}{codigo ibge inventado}
  \item{valor}{valor do fundo}
  \item{alunos}{numero de alunos por ente federativo}
}}
\source{
Tabela Criada
}
\usage{
test
}
\description{
Uma base de dados para testar as funcoes do pacote
}
\keyword{datasets}
