% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teste_simulador.R
\name{testa_dados_simulador}
\alias{testa_dados_simulador}
\title{Testa dados}
\usage{
testa_dados_simulador(
  dados_matriculas,
  dados_complementar,
  dados_peso,
  complementacao_vaaf,
  complementacao_vaat,
  complementacao_vaar,
  teto,
  chao
)
}
\arguments{
\item{dados_matriculas}{data.frame com os dados de matriculas}

\item{dados_complementar}{data.frame com os dados de peso das etapas}

\item{dados_peso}{data.frame com os dados de peso das etapas}

\item{complementacao_vaaf}{valor numerico com o montante a ser complementado pela uniao na etapa VAAF}

\item{complementacao_vaat}{valor numerico com o montante a ser complementado pela uniao na etapa VAAT}

\item{complementacao_vaar}{valor numerico com o montante a ser complementado pela uniao na etapa VAAR}
}
\description{
Checa se os objetos usados na simulacao sao adequados para realiza-la
}
