% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{teste_complementar}
\alias{teste_complementar}
\title{Dados Teste}
\format{
Uma tabela com 3 colunas e 29 linhas.
\describe{
  \item{ibge}{Codigo ibge do ente}
  \item{uf}{Variavel em caractere com o nome das unidades da federacao}
  \item{nome}{Variavel em caracteres com o nome do ente}
  \item{recursos_vaaf}{Receitas do fundeb cada ente federativo utilizadas na etapa VAAF}
  \item{recursos_vaat}{Receitas do fundeb cada ente federativo utilizadas na etapa VAAT}
  \item{nse}{Variaval com o indice socioeconomico, no exemplo do teste todos os valores sao um}
  \item{nf}{Variaval com o indice de nivel fiscal, no exemplo do teste todos os valores sao um}
  \item{peso_vaar}{Parcela dos recursos vaar que irao para cada ente}
  \item{inabilitados_vaat}{Variavel com o identificador dos entes que nao participarao da complementacao vaat}
  }
}
\usage{
teste_complementar
}
\description{
Uma base de dados com as matriculas por etapa, coeficiente de distribuicao, estimativas de receitas e estado por ente federativo
}
\keyword{datasets}
