% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{teste_peso}
\alias{teste_peso}
\title{Peso}
\format{
Uma tabela com 3 colunas e 29 linhas.
\describe{
  \item{etapa}{Variavel em caractere com o nome das etaoas segundo as portarias do FNDE}
  \item{peso_vaaf}{Variavel numerica com o peso de cada etapa para a etapa de complementacao VAAF segunda o regulamentacao do FNDE}
  \item{peso_vaat}{Variavel numerica com o peso de cada etapa para a etapa de complementacao VAAT}
  \item{encontrado}{Variavel binaria, 1 se encontrado na regulamentacao do FNDE e 0 caso contrario}
}
}
\source{
Tabela Criada a partir dos dados do TCU
}
\usage{
teste_peso
}
\description{
Uma base de dados com os pesos por etapa
}
\keyword{datasets}
