% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une-vaaf.R
\name{une_vaaf}
\alias{une_vaaf}
\title{Une a equalizacao vaaf com a tabela de entes}
\usage{
une_vaaf(dados_entes, dados_etapa, produto_dt = TRUE)
}
\arguments{
\item{dados_entes}{data.frame ou data.table com os dados dos entes federativos}

\item{produto_dt}{Variavel binaria que determina se a tabela produto da funcao sera da classe data.table (se Verdadeiro) ou data.frame (se Falso)}

\item{dados_fundo}{data.frame ou data.table com os dados da equalizacao da etapa}

\item{etapa}{variavel em caracteres que pode ter os valores "vaat" ou "vaaf" a depender da etapa do calculo}
}
\value{
Um data.frame ou data.table com a simulacao dos dados do FNDE
}
\description{
Une a equalização do fundo etapa vaaf com a tabela dos entes e calcula o valor aluno ano da etapa
}
