% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alunos-2015.R
\docType{data}
\name{alunos_2015}
\alias{alunos_2015}
\title{Base com dados de alunos}
\format{Uma tabela com 6 colunas e 145264 linhas.
\describe{
  \item{ibge}{codigo ibge do ente federativo}
  \item{nome}{nome do ente federativo}
  \item{regiao}{regiao do ente federativo}
  \item{estado}{estado do ente federativo}
  \item{etapa}{etapa de ensino}
  \item{alunos}{numero de alunos por etapa de ensino}
}}
\source{
Tabela provida pela organizacao Todos pela educacao
}
\usage{
alunos_2015
}
\description{
Base com dados do numero de alunos por ente federativo para o ano de 2015
}
\keyword{datasets}
