% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula-alunos-estadual.R
\name{calcula_alunos_estadual}
\alias{calcula_alunos_estadual}
\title{Calcula numero de alunos por estado}
\usage{
calcula_alunos_estadual(dados, codigo = ibge, var_alunos = alunos)
}
\arguments{
\item{dados}{data.frame com o numero de alunos por ente federativo}

\item{codigo}{coluna numerica com o codigo identificador do ente}

\item{var_alunos}{nomeia a coluna numerica com o numero de alunos por ente}
}
\value{
Um data.frame com o numero de alunos por estado
}
\description{
Calcula o numero de alunos por ente federativo. Essa funcao supoe que os dois primeiros digitos sao referentes ao estado do ente federativo
}
\examples{
library(simulador.fundeb)
simulador.fundeb:::calcula_alunos_estadual(test)
}
