% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcula-fundo-estadual.R
\name{calcula_fundo_estadual}
\alias{calcula_fundo_estadual}
\title{Calcula valor do fundo total}
\usage{
calcula_fundo_estadual(fundo, codigo = ibge, var_fundo = fundeb)
}
\arguments{
\item{fundo}{data.frame com o valor do fundo de cada entede federativo e o codigo ibge do ente}

\item{codigo}{coluna numerica com o codigo do estado}

\item{var_fundo}{coluna numerica com o valor do fundo}
}
\value{
Um data.frame com o valor do fundo estadual em um coluna e o codigo do respectivo estado em outra
}
\description{
Essa função calcula o valor total de um fundo
}
\examples{
library(simulador.fundeb)
simulador.fundeb:::calcula_fundo_estadual(test)
}
