% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{complementar_teste}
\alias{complementar_teste}
\title{Dados Teste}
\format{
Uma tabela com 3 colunas e 29 linhas.
\describe{
  \item{ibge}{Codigo ibge do ente}
  \item{impostos_extra}{Impostos extra com direcionamento a educacao}
  \item{impostos_cap}{Impostos extra por pessoa}
  \item{nome}{Variavel em caracteres com o nome do ente}
  \item{uf}{Variavel em caractere com o nome das unidades da federacao}
  \item{idhm}{Variaval com o indice de desenvolvimento humano do ente}
  }
}
\usage{
complementar_teste
}
\description{
Uma base de dados com as matriculas por etapa, coeficiente de distribuicao, estimativas de receitas e estado por ente federativo
}
\keyword{datasets}
