% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\name{complementar_teste}
\alias{complementar_teste}
\title{Dados Teste}
\format{
Uma tabela com 3 colunas e 29 linhas.
\describe{
  \item{ibge}{Codigo ibge do ente}
  \item{recursos_extra}{Variavel com recursos extra com direcionamento a educacao}
  \item{fator_fiscal}{Variaval com Impostos extra por pessoa que sera usada como fator fiscal}
  \item{nome}{Variavel em caracteres com o nome do ente}
  \item{uf}{Variavel em caractere com o nome das unidades da federacao}
  \item{fator_social}{Variaval com o indice de desenvolvimento humano do ente que sera usado como fator social}
  \item{fundeb_vaaf}{Receitas do fundeb cada ente federativo utilizadas na etapa VAAF}
  \item{fundeb_vaat}{Receitas do fundeb cada ente federativo utilizadas na etapa VAAT}
  }
}
\usage{
complementar_teste
}
\description{
Uma base de dados com as matriculas por etapa, coeficiente de distribuicao, estimativas de receitas e estado por ente federativo
}
\keyword{datasets}
