% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equaliza_fundo.R
\name{equaliza_fundo}
\alias{equaliza_fundo}
\title{Equaliza fundo}
\usage{
equaliza_fundo(
  dados,
  complementacao_uniao,
  var_ordem,
  var_alunos,
  var_recursos,
  identificador,
  entes_excluidos = NULL
)
}
\arguments{
\item{dados}{Um objeto da classe data.frame com os dados do fundo a serem equalizados}

\item{complementacao_uniao}{Numero com a quantidade de fundos a serem utilizados na equalizacao}

\item{var_ordem}{Variavel que sera utilizada para ordenar os entes}

\item{var_alunos}{Variavel que sera utilizada como numero de alunos a serem considerados do ente}

\item{var_recursos}{Variavel com as receitas que seram equalizadas}

\item{identificador}{Identificador unico para cada ente que estara sendo equalizado}

\item{entes_excluidos}{Vetor com o identificador de cada entre que sera removido da equalizacao}
}
\value{
Um data.frame
}
\description{
Recebe dados dos entes do fnde, divididos em entes federativos ou fundos estaduais, e equaliza o valor recebido de acordo com o vaa em questao. A equalizacao e feita de baixo para cima.
}
