% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equaliza-modelo.R
\name{equaliza_modelo}
\alias{equaliza_modelo}
\title{Equaliza um modelo}
\usage{
equaliza_modelo(dados,
  var_necessidade_equalizacao = necessario_equalizacao,
  aporte = aporte_federal, fundo_equalizado = fundo_estadual,
  var_alunos = alunos_estado, codigo)
}
\arguments{
\item{dados}{data.frame que serao equalizados}

\item{var_necessidade_equalizacao}{variavel que indica a necessidade de equalizacao}

\item{aporte}{quantidade de verba que será usada na equalização}

\item{fundo_equalizado}{coluna numerica com o fundo a ser equalizado entre os entes}
}
\value{
Um data.frame com as seguintes colunas adicionas cumulativo de alunos e gasto necessario para equalizar
}
\description{
Função intermediária que equaliza o valor do fundo por aluno
}
\examples{
library(simulador.fundeb)
}
