% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gera-dados-estaduais.R
\name{gera_dados_estaduais}
\alias{gera_dados_estaduais}
\title{Gera tabela com dados dos estados}
\usage{
gera_dados_estaduais(dados)
}
\arguments{
\item{dados}{data.frame preparado por pondera_geral}
}
\value{
Um data.frame com as seguintes colunas adicionas cumulativo de alunos e gasto necessario para equalizar
}
\description{
Função intermediária que gera dados dos estados que será usada nas simulações
}
\examples{
library(simulador.fundeb)
}
