% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera_alunos_etapa.R
\name{pondera_alunos_etapa}
\alias{pondera_alunos_etapa}
\title{Pondera alunos por etapa}
\usage{
pondera_alunos_etapa(dados_alunos, dados_peso)
}
\arguments{
\item{dados_alunos}{data.frame com os dados de alunos}

\item{dados_peso}{data.frame com os dados de peso das etapas}
}
\value{
Um data.frame ou data.table com os dados de alunos considerando os pesos dados de FNDE em formato longo ou curto
}
\description{
Pondera dados de alunos do FNDE por ente e etapa e retorna os valores ponderados por etapa
}
