% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-alunos-etapa.R
\name{pondera_alunos_etapa}
\alias{pondera_alunos_etapa}
\title{Pondera alunos por etapa}
\usage{
pondera_alunos_etapa(
  dados_fnde,
  peso_etapas = peso,
  difere_etapas_complementacao = c("vaaf_vaat", "mesmos_pesos"),
  produto_dt = TRUE
)
}
\arguments{
\item{peso_etapas}{data.frame com os dados de peso das etapas}

\item{difere_etapas_complementacao}{variavel em caractere com as opcoes de diferenciacao dos pesos das etapas. Caso escolha-se vaaf_vaat a etapa vaaf e a etapa vaat consideram pesos diferentes, caso escolha-se mesmos pesos as etapas tem os mesmos pesos}

\item{produto_dt}{valor logico que determina se o resultado sera em data.table (caso produto_dt = TRUE) ou data.frame (caso produto_dt = FALSE)}
}
\value{
Um data.frame ou data.table com os dados de alunos considerando os pesos dados de FNDE em formato longo ou curto
}
\description{
Pondera dados de alunos do FNDE por ente e etapa e retorna os valores ponderados por etapa
}
