% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-alunos-rede.R
\name{pondera_alunos_rede}
\alias{pondera_alunos_rede}
\title{Calcula valor do fundo total}
\usage{
pondera_alunos_rede(
  base_alunos,
  codigo = ibge,
  etapa_ensino = etapa,
  var_alunos = alunos,
  busca_creche = "Creche|Pré-escola",
  busca_em = "EM"
)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{codigo}{coluna numerica com o codigo do estado}

\item{etapa_ensino}{coluna de fator ou caractere com a etapa de ensino}

\item{var_alunos}{coluna numerica com o numero de alunos}

\item{busca_creche}{caractere com expressao que identifica a etapa referente a creche}

\item{busca_em}{caractere com expressao que identifica a etapa referente ao ensino medio}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com código ibge, alunos e
}
\examples{
library(simulador.fundeb)

}
