% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-geral.R
\name{pondera_geral}
\alias{pondera_geral}
\title{Pondera alunos por etapa de ensino}
\usage{
pondera_geral(base_alunos, ponderador, base_socioeconomica,
  condicao_rede = TRUE, condicao_etapa = TRUE,
  condicao_socio = FALSE)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{ponderador}{data.frame de peso de aluno por etapa}

\item{codigo}{coluna numerica com o codigo do estado}

\item{etapa_ensino}{coluna de fator ou caractere com a etapa de ensino}

\item{var_alunos}{coluna numerica com o numero de alunos}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com código ibge, etapa e alunos e outra com pesos de ponderador por etapa e gera uma tabela com o número de alunos ponderados por entidade da federação
}
\examples{
library(simulador.fundeb)
}
