% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-geral.R
\name{pondera_geral}
\alias{pondera_geral}
\title{Pondera alunos por etapa de ensino}
\usage{
pondera_geral(
  base_alunos,
  ponderador,
  base_socioeconomica,
  base_financas,
  condicao_rede = TRUE,
  condicao_socio = TRUE,
  min_social = 1,
  max_social = 1.3,
  min_financas = 1,
  max_financas = 1.3,
  var_socioeconomica = nse,
  considerar = "ambos",
  ...
)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{base_socioeconomica}{data.frame o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame o codigo do ibge e valores do fundeb e das demais receitas}

\item{condicao_rede}{valor binario que condiciona a ponderacao por rede}

\item{min_social}{peso minimo dado a informacao socioeconomica}

\item{max_social}{peso maximo dado a informacao socioeconomica}

\item{min_financas}{peso minimo dado a informacao de financas}

\item{max_financas}{peso maximo dado a informacao de financas}

\item{var_socioeconomica}{variavel numerica socioeconomica de um ente federativo}

\item{considerar}{nome que indica que parametro sera considarado na criacao do peso socioeconomico, se selecionado social se considerara apenas as informacoes sociais, se selecionado financas se considerara apenas as informacoes de financas e se selecionado ambos se considerara ambas as dimensoes}

\item{condicao_etapa}{valor binario que condiciona a ponderacao por etapa}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com codigo ibge, etapa e alunos e outra com pesos de ponderador por etapa e gera uma tabela com o número de alunos ponderados por entidade da federação
}
\examples{
library(simulador.fundeb)
}
