% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-socioeconomico.R
\name{pondera_socioeconomico}
\alias{pondera_socioeconomico}
\title{Pondera alunos por etapa de ensino}
\usage{
pondera_socioeconomico(base_socioeconomica,
  variavel_socioeconomica = nse, min_social = 1, max_social = 1.3)
}
\arguments{
\item{base_socioeconomica}{data.frame o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{variavel_socioeconomica}{variavel numerica socioeconomica de um ente federativo}

\item{min_social}{peso minimo dado a informacao socioeconomica}

\item{max_social}{peso maximo dado a informacao socioeconomica}

\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com código ibge, etapa e alunos e outra com pesos de ponderador por etapa e gera uma tabela com o número de alunos ponderados por entidade da federação
}
\examples{
library(simulador.fundeb)

}
