% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pondera-socioeconomico.R
\name{pondera_socioeconomico}
\alias{pondera_socioeconomico}
\title{Pondera alunos por etapa de ensino}
\usage{
pondera_socioeconomico(
  base_alunos,
  base_socioeconomica,
  base_financas,
  var_socioeconomica = nse,
  var_fundo_pond = demais_receitas,
  min_social = 1,
  max_social = 1.3,
  codigo = ibge
)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{base_socioeconomica}{data.frame o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame o codigo do ibge e valores do fundeb e das demais receitas}

\item{var_socioeconomica}{variavel numerica socioeconomica de um ente federativo}

\item{var_fundo_pond}{variavel numerica com o valor do fundo a ser considerado na ponderacao financeira}

\item{min_social}{peso minimo dado a informacao socioeconomica}

\item{max_social}{peso maximo dado a informacao socioeconomica}

\item{codigo}{parametro com o nome do codigo a ser usado como identificador dos entes federativos, recomedanda-se o codigo ibge}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com código ibge, etapa e alunos e outra com pesos de ponderador por etapa e gera uma tabela com o número de alunos ponderados por entidade da federação
}
\examples{
library(simulador.fundeb)

}
