% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ponderador-alunos.R
\docType{data}
\name{ponderador_alunos}
\alias{ponderador_alunos}
\title{Base de dados com pesos das etapas para ponderacao do numero de alunos}
\format{Uma tabela com 2 colunas e 28 linhas.
\describe{
  \item{etapa}{etapa de ensino}
  \item{ponderador}{peso do aluno por etapa}
}}
\source{
Tabela provida pela organizacao Todos pela educacao
}
\usage{
ponderador_alunos
}
\description{
Base de dados com pesos das etapas para ponderacao do numero de alunos para realizacao das simulacoes do financiamento
}
\keyword{datasets}
