% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simula-fnde.R
\name{simula_fundeb}
\alias{simula_fundeb}
\title{Simula FNDE}
\usage{
simula_fundeb(dados_fnde, peso_etapas = peso, aporte, produto_dt = TRUE)
}
\arguments{
\item{dados_fnde}{data.frame ou data.table com os dados do FNDE}

\item{peso_etapas}{data.frame com os dados de peso das etapas}

\item{produto_dt}{Variavel binaria que determina se a tabela produto da funcao sera da classe data.table (se Verdadeiro) ou data.frame (se Falso)}
}
\value{
Um data.frame ou data.table com a simulacao dos dados do FNDE
}
\description{
Simula a distribuição dos fundos do FNDE e as etapas de equalização
}
