% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simula_fundeb.R
\name{simula_fundeb}
\alias{simula_fundeb}
\title{Simula Fundeb}
\usage{
simula_fundeb(
  dados_matriculas,
  dados_complementar,
  dados_peso,
  max_nse = 1.05,
  min_nse = 0.95,
  max_nf = 1.05,
  min_nf = 0.95,
  complementacao_vaaf,
  complementacao_vaat,
  complementacao_vaar,
  checa_valores = FALSE
)
}
\arguments{
\item{dados_matriculas}{data.frame com os dados de matriculas}

\item{dados_complementar}{data.frame com os dados de peso das etapas}

\item{dados_peso}{data.frame com os dados de peso das etapas}

\item{max_nse}{valor maximo do nivel socioeconomico na ponderacao de matriculas}

\item{min_nse}{valor minimo do nivel socioeconomico na ponderacao de matriculas}

\item{max_nf}{valor maximo do nivel de capacidade fiscal na ponderacao de matriculas, exclusivo para VAAF}

\item{min_nf}{valor minimo do nivel de capacidade fiscal na ponderacao de matriculas, exclusivo para VAAF}

\item{complementacao_vaaf}{valor numerico com o montante a ser complementado pela uniao na etapa VAAF}

\item{complementacao_vaat}{valor numerico com o montante a ser complementado pela uniao na etapa VAAT}

\item{complementacao_vaar}{valor numerico com o montante a ser complementado pela uniao na etapa VAAR}

\item{checa_valores}{valor logico para avaliacao da necessidade de checar valores}
}
\value{
Um data.frame ou com a simulacao dos dados do FNDE
}
\description{
Simula a distribuição dos fundos do Fundeb e as etapas de equalização
}
\examples{

data("teste_matriculas")
data("teste_complementar")
data("teste_peso")

df_teste = simulador.fundeb::simula_fundeb(
  dados_matriculas = teste_matriculas,
  dados_complementar = teste_complementar,
  dados_peso = teste_peso,
  complementacao_vaaf = 4e5,
  complementacao_vaat = 1e5,
  complementacao_vaar = 1e5)

head(df_teste)

}
