% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simula_fundeb.R
\name{simula_fundeb}
\alias{simula_fundeb}
\title{Simula Fundeb}
\usage{
simula_fundeb(
  dados_alunos,
  dados_complementar,
  dados_peso,
  teto = 1.05,
  chao = 0.95,
  complementacao_vaaf,
  complementacao_vaat,
  complementacao_vaar
)
}
\arguments{
\item{dados_alunos}{data.frame com os dados de alunos}

\item{dados_complementar}{data.frame com os dados de peso das etapas}

\item{dados_peso}{data.frame com os dados de peso das etapas}

\item{teto}{valor maximo do nivel socioeconomico na ponderacao de alunos}

\item{chao}{valor minimo do nivel socioeconomico na ponderacao de alunos}

\item{complementacao_vaaf}{valor numerico com o montante a ser complementado pela uniao na etapa VAAF}

\item{complementacao_vaat}{valor numerico com o montante a ser complementado pela uniao na etapa VAAT}

\item{complementacao_vaar}{valor numerico com o montante a ser complementado pela uniao na etapa VAAR}
}
\value{
Um data.frame ou data.table com a simulacao dos dados do FNDE
}
\description{
Simula a distribuição dos fundos do Fundeb e as etapas de equalização
}
