% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simula-fnde.R
\name{simula_fundeb}
\alias{simula_fundeb}
\title{Simula FNDE}
\usage{
simula_fundeb(
  dados_fnde,
  dados_complementar,
  peso_etapas = peso,
  chao_socio = 1,
  teto_socio = 1.3,
  chao_fiscal = 1,
  teto_fiscal = 1.3,
  entes_excluidos_vaaf = NULL,
  complementacao_vaaf,
  complementacao_vaat,
  produto_dt = TRUE
)
}
\arguments{
\item{dados_complementar}{data.frame com os dados de peso das etapas}

\item{peso_etapas}{data.frame com os dados de peso das etapas}

\item{chao_socio}{valor minimo do peso da variavel social na ponderacao de alunos}

\item{teto_socio}{valor maximo do peso da variavel social na ponderacao de alunos}

\item{chao_fiscal}{valor minimo do peso da variavel fiscal na ponderacao de alunos}

\item{teto_fiscal}{valor maximo do peso da variavel fiscal na ponderacao de alunos}

\item{complementacao_vaaf}{valor numerico com o montante a ser complementado pela uniao na etapa VAAF}

\item{complementacao_vaat}{valor numerico com o montante a ser complementado pela uniao na etapa VAAT}
}
\value{
Um data.frame ou data.table com a simulacao dos dados do FNDE
}
\description{
Simula a distribuição dos fundos do FNDE e as etapas de equalização
}
