% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelo-hibrido.R
\name{simular_modelo_hibrido}
\alias{simular_modelo_hibrido}
\title{Simulação do modelo híbrido de financiamento da educação}
\usage{
simular_modelo_hibrido(base_alunos, ponderador, base_socioeconomica,
  base_financas, auxilio_federal_fundeb = 0.1,
  auxilio_federal_vaa = 0.05, var_fundo = fundeb,
  var_alunos = alunos, ...)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{ponderador}{data.frame de peso de aluno por etapa}

\item{base_socioeconomica}{data.frame com  o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame com o codigo do ibge e os valores do fundeb e das demais receitas}

\item{auxilio_federal_fundeb}{percentual do fundo que a União complementará segundo o modelo fundeb}

\item{auxilio_federal_vaa}{percentual do fundo que a União complementará segundo o modelo vaa}

\item{var_fundo}{parametro com o nome da variavel do fundo a ser considerado}

\item{var_alunos}{parametro com o nome da variavel alunos a ser considerada}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com numero de alunos por ente e por etapa, ponderador por etapa, dados socioeconomicos por ente e dados financeiros por ente e simula o modelo híbrido de financiamento da educação
}
