% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simular-modelo-hibrido-tempo.R
\name{simular_modelo_hibrido_tempo}
\alias{simular_modelo_hibrido_tempo}
\title{Simulação do modelo hibrido de financiamento da educação no tempo}
\usage{
simular_modelo_hibrido_tempo(
  base_alunos,
  ponderador,
  base_socioeconomica,
  base_financas,
  auxilio_federal = 0.1,
  auxilio_federal_vaa = 0.05,
  equalizacao_socio = FALSE,
  distribuicao_fundo_estadual_socio = FALSE,
  crescimento_economico,
  crescimento_demografico,
  ...
)
}
\arguments{
\item{base_alunos}{data.frame de numero de alunos por etapa e ente federativo}

\item{ponderador}{data.frame de peso de aluno por etapa}

\item{base_socioeconomica}{data.frame com  o codigo do ibge, o numero de alunos ponderado e uma variavel socioeconomia a ponderar}

\item{base_financas}{data.frame com o codigo do ibge e os valores do fundeb e das demais receitas}

\item{auxilio_federal}{percentual do fundo que a União complementará}

\item{auxilio_federal_vaa}{percentual do fundo que a União complementará segundo o modelo vaa}

\item{equalizacao_socio}{parametro lógico que controla se a equalização do fundo considerara o vetor de alunos ou de alunos socioeconomico}

\item{distribuicao_fundo_estadual_socio}{parametro logico que controla se a distribuicao do fundo estadual considerara o vetor de alunos ou de alunos socioeconomico}

\item{crescimento_economico}{vetor númerico de crescimento economico, especificamente do fundeb e das demais receitas}

\item{crescimento_demografico}{vetor númerico de crescimento demográfico para alunos}
}
\value{
Data.frame com alunos ponderador por ente federativo
}
\description{
Recebe uma base com numero de alunos por ente e por etapa, ponderador por etapa, dados socioeconomicos por ente e dados financeiros por ente e simula o modelo fundeb de financiamento da educação ao longo do tempo. Considera o crescimento demográfico e econômico ao longo do tempo.
}
