% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une_vaaf.R
\name{une_vaaf}
\alias{une_vaaf}
\title{Une a equalizacao vaaf com a tabela de entes}
\usage{
une_vaaf(dados_entes, dados_estados, dados_fundos_estaduais)
}
\arguments{
\item{dados_entes}{data.frame com os dados dos entes federativos considerados na funcao}

\item{dados_estados}{data.frame com os dados dos fundos estaduais}

\item{dados_fundos_estaduais}{data.frame com os dados da equalizacao da etapa dos fundos estaduais}
}
\value{
Um data.frame ou data.table com a simulacao dos dados do FNDE
}
\description{
Une a equalização do fundo etapa vaaf com a tabela dos entes e calcula o valor aluno ano da etapa
}
