% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/une-vaat.R
\name{une_vaat}
\alias{une_vaat}
\title{Une a equalizacao vaat com a tabela de entes}
\usage{
une_vaat(dados_entes, dados_etapa, produto_dt = TRUE)
}
\arguments{
\item{dados_entes}{data.frame ou data.table com os dados dos entes federativos}

\item{dados_etapa}{data.frame ou data.table com os dados da equalizacao da etapa}
}
\value{
Um data.frame ou data.table com a simulacao dos dados do FNDE
}
\description{
Une a equalização do fundo etapa vaat com a tabela dos entes e calcula o valor aluno ano da etapa

Une a equalização do fundo com a tabela dos entes e calcula o valor aluno ano da etapa
}
